Backend/requirements.txt
flask
Flask-Cors
python-dotenv
requests
supabase
sendgrid
twilio
gunicorn